# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AI Website API Documentation",
        "description": "API documentation for AI Website backend service",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/vendors/{vendor}/products": {
            "get": {
                "tags": [
                    "Vendor - Products"
                ],
                "summary": "List all products belonging to the specified vendor",
                "operationId": "122a6d3afd0efe49b0d5333ec27fa069",
                "parameters": [
                    {
                        "name": "vendor",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of vendor products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/App\\Http\\Resources\\ProductResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Vendor - Products"
                ],
                "summary": "Create a new product by vendor",
                "operationId": "709529ad0827ecdc2599c9d6759ca241",
                "parameters": [
                    {
                        "name": "vendor",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/App\\Http\\Requests\\Vendor\\StoreProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/App\\Http\\Resources\\ProductResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/vendors/{vendor}/products/{id}": {
            "put": {
                "tags": [
                    "Vendor - Products"
                ],
                "summary": "Update existing product",
                "operationId": "0fb65348a0bab3eed91e2cd3950070bb",
                "parameters": [
                    {
                        "name": "vendor",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/App\\Http\\Requests\\Vendor\\UpdateProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/App\\Http\\Resources\\ProductResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Vendor - Products"
                ],
                "summary": "Delete product",
                "operationId": "d6c914e5925f34697c37c37e5b79f041",
                "parameters": [
                    {
                        "name": "vendor",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted successfully"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/vendors/{vendor}/products/{id}/files": {
            "post": {
                "tags": [
                    "Vendor - Products"
                ],
                "summary": "Upload digital file for product",
                "operationId": "70a99cd8539cb5fa6c2c69cb7667929e",
                "parameters": [
                    {
                        "name": "vendor",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/App\\Http\\Requests\\Vendor\\UploadProductFileRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "File uploaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/App\\Http\\Resources\\ProductFileResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/vendors/{vendor}/products/{id}/license-keys": {
            "post": {
                "tags": [
                    "Vendor - Products"
                ],
                "summary": "Import license keys into product pool",
                "operationId": "32ad0ddb9b862d50e1b43cbdc5412f32",
                "parameters": [
                    {
                        "name": "vendor",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/App\\Http\\Requests\\Vendor\\ImportLicenseKeysRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "License keys imported successfully"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}